% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fla_app.R
\name{fla_app}
\alias{fla_app}
\title{Lance l'application dashboard de Flavia APE}
\usage{
fla_app()
}
\value{
Lance l'application dashboard de Flavia APE
}
\description{
Lance l'application dashboard de Flavia APE
}
