% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fla_generate_atlas.R
\name{fla_generate_atlas}
\alias{fla_generate_atlas}
\title{Génère un lot de cartes exportées au format .png (utilise en boucle la fonction fla_generate_map)}
\usage{
fla_generate_atlas(
  df = data_sp,
  territory = NULL,
  url_basemap = NULL,
  wkt_field = "wkt",
  rotate_field = NULL,
  label_field = NULL,
  data_color = "#ff7f00",
  fillPolygons = 1,
  attrib = NULL,
  scale = F
)
}
\arguments{
\item{df}{Dataframe contenant les données spatiales à cartographier, dont un champs contenant la géométrie au format WKT (projection WGS84)}

\item{territory}{Objet comportant la géométrie du territoire (geojson chargé avec st_read par exemple). Peut comporter un ou plusieurs polygones.}

\item{url_basemap}{URL du service WMS fournissant le fond de carte. Par défaut, c'est le fond topographique de l'IGN qui est utilisé : https://wxs.ign.fr/inspire/inspire/r/wms}

\item{wkt_field}{Chaîne de caractères décrivant le nom du champs contenant la géométrie WKT dans le dataframe source. (par défaut, "wkt")}

\item{rotate_field}{Chaîne de caractères décrivant le nom du champs à utiliser pour les itérations de l'atlas}

\item{label_field}{Chaîne de caractères décrivant le nom du champs décrivant l'étiquette de chaque carte}

\item{data_color}{Chaîne de caractères permettant de définir la couleur d'affichage des données (sous forme de variable "blue" ou de code hexadécimal "#ff7f00"). (Par défaut, "#ff7f00")}

\item{fillPolygons}{Paramètre permettant de remplir (1) ou non (0) les géométries des données "au polygone", ou de varier l'opacité du remplissage (ex. 0.5)}

\item{attrib}{Chaîne de caractères permettant d'afficher l'attribution en pied de carte (par défaut : "Flavia APE - 2024")}

\item{scale}{Paramètre TRUE/FALSE permettant d'afficher une échelle sous la carte (par défaut : TRUE)}
}
\value{
Un lot de cartes au format .png, exportées au sein d'un dossier "atlas" dans l'environnement de travail. Chaque fichier est nommé "valeur_du_champ_d'itération.png"
}
\description{
Génère un lot de cartes exportées au format .png (utilise en boucle la fonction fla_generate_map)
}
