% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fla_generate_map.R
\name{fla_generate_map}
\alias{fla_generate_map}
\title{Cartographie standardisée des données spatiales.}
\usage{
fla_generate_map(
  df = NULL,
  territory = NULL,
  url_basemap = NULL,
  wkt_field = "wkt",
  data_color = "#ff7f00",
  fillPolygons = 1,
  title = NULL,
  attrib = "Flavia APE - 2024",
  scale = T,
  print = T,
  filename = "map.png"
)
}
\arguments{
\item{df}{Dataframe contenant les données spatiales à cartographier, dont un champs contenant la géométrie au format WKT (projection WGS84)}

\item{territory}{Objet comportant la géométrie du territoire (geojson chargé avec st_read par exemple). Peut comporter un ou plusieurs polygones.}

\item{url_basemap}{URL du service WMS fournissant le fond de carte. Par défaut, c'est le fond topographique de l'IGN qui est utilisé : https://wxs.ign.fr/inspire/inspire/r/wms}

\item{wkt_field}{Chaîne de caractères décrivant le nom du champs contenant la géométrie WKT dans le dataframe source. (par défaut, "wkt")}

\item{data_color}{Chaîne de caractères permettant de définir la couleur d'affichage des données (sous forme de variable "blue" ou de code hexadécimal "#ff7f00"). (Par défaut, "#ff7f00")}

\item{fillPolygons}{Paramètre permettant de remplir (1) ou non (0) les géométries des données "au polygone", ou de varier l'opacité du remplissage (ex. 0.5)}

\item{title}{Chaîne de caractères permettant d'afficher un titre en tête de la carte.}

\item{attrib}{Chaîne de caractères permettant d'afficher l'attribution en pied de carte (par défaut : "Flavia APE - 2024")}

\item{scale}{Paramètre TRUE/FALSE permettant d'afficher une échelle sous la carte (par défaut : TRUE)}

\item{print}{Paramètre TRUE/FALSE permettant d'exporter la carte dans le répertoire de travail au format image .png. (Par défaut: TRUE)}

\item{filename}{Chaîne de caractère définissant le nom du fichier png si Print=T. (Par défaut : "map.png")}
}
\value{
La carte des données spatiales, avec un masque et une emprise correspondant au territoire fourni.
}
\description{
Cartographie standardisée des données spatiales.
}
