% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fla_get_geom_type.R
\name{fla_get_geom_type}
\alias{fla_get_geom_type}
\title{Calcule la diversité biologique inventoriée dans un lot de données}
\usage{
fla_get_geom_type(df, wkt_field)
}
\arguments{
\item{wkt_field}{Le nom du champs WKT à utiliser. Doit contenir un WKT en WGS84 (4326)}

\item{dt}{La table de données comportant les données à cartographier}
}
\value{
Renvoie la table de données avec un champs supplémentaire contenant la validité et le type de la géométrie WKT.
}
\description{
Calcule la diversité biologique inventoriée dans un lot de données
}
