% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fla_nb_jdd.R
\name{fla_nb_jdd}
\alias{fla_nb_jdd}
\title{Calcule le nombre de jeux de données distincts dans un lot de données}
\usage{
fla_nb_jdd(data, column = "jdd_uuid")
}
\arguments{
\item{data}{La table de données importées depuis GeoNature}

\item{column}{Le nom du champs à utiliser pour calculer le nombre de jeux de données (jdd_uuid par défaut)}
}
\value{
Le nombre de taxons compris dans un lot de données
}
\description{
Calcule le nombre de jeux de données distincts dans un lot de données
}
