% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fla_nb_obs.R
\name{fla_nb_obs}
\alias{fla_nb_obs}
\title{Calcule le nombre d'observations dans un lot de données}
\usage{
fla_nb_obs(data)
}
\arguments{
\item{data}{La table de données importées depuis GeoNature}
}
\value{
Le nombre d'observations dans un lot de données
}
\description{
Calcule le nombre d'observations dans un lot de données
}
