% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fla_nb_releves.R
\name{fla_nb_releves}
\alias{fla_nb_releves}
\title{Calcule le nombre de relevés effectués au sein d'un lot de données.
Basé sur le nombre de combinaisons différentes entre les date_min,
date_max et localisation des observations.}
\usage{
fla_nb_releves(
  data,
  date_debut = "date_debut",
  date_fin = "date_fin",
  localisation = "wkt_l93"
)
}
\arguments{
\item{data}{La table de données importées depuis GeoNature}

\item{date_debut}{Le nom du champs à utiliser pour la date de début de la session (date_debut par défaut)}

\item{date_fin}{Le nom du champs à utiliser pour la date de fin de la session (date_fin par défaut)}

\item{localisation}{Le nom du champs à utiliser pour la localisation du relevé (wkt_l93 par défaut)}
}
\value{
Le nombre de relevés effectués dans un lot de données
}
\description{
Calcule le nombre de relevés effectués au sein d'un lot de données.
Basé sur le nombre de combinaisons différentes entre les date_min,
date_max et localisation des observations.
}
