% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fla_nb_sessions.R
\name{fla_nb_session}
\alias{fla_nb_session}
\title{Calcule le nombre de sessions de terrain au sein d'un lot de données.
Basé sur les combinaisons date_min, date_max des observations. Une chasse
de jour et une chasse de nuit compteront donc pour deux sessions (une diurne
et une nocturne)}
\usage{
fla_nb_session(data, date_debut = "date_debut", date_fin = "date_fin")
}
\arguments{
\item{data}{La table de données importées depuis GeoNature}

\item{date_debut}{Le nom du champs à utiliser pour la date de début de la session (date_debut par défaut)}

\item{date_fin}{Le nom du champs à utiliser pour la date de fin de la session (date_fin par défaut)}
}
\value{
Le nombre de taxons compris dans un lot de données
}
\description{
Calcule le nombre de sessions de terrain au sein d'un lot de données.
Basé sur les combinaisons date_min, date_max des observations. Une chasse
de jour et une chasse de nuit compteront donc pour deux sessions (une diurne
et une nocturne)
}
