% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fla_rapport.R
\name{fla_rapport}
\alias{fla_rapport}
\title{Synthese globale d'un lot de données
Fonction permettant d'obtenir les informations de synthèse sur un lot de données
Renvoie le nombre de données, de taxons, de relevés, les méthodes utilisées etc.}
\usage{
fla_rapport(data)
}
\arguments{
\item{data}{Le tableau R des données à analyser à partir d'un export GeoNature}
}
\value{
Rapport de synthèse du lot de données
}
\description{
Synthese globale d'un lot de données
Fonction permettant d'obtenir les informations de synthèse sur un lot de données
Renvoie le nombre de données, de taxons, de relevés, les méthodes utilisées etc.
}
