% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_from_geonature.R
\name{load_from_geonature}
\alias{load_from_geonature}
\title{Importer un fichier csv provenant de GeoNature
Une fonction pour importer un fichier csv issu d'un export
GeoNature, sans modification d'encodage ni de format.}
\usage{
load_from_geonature(file)
}
\arguments{
\item{file}{Chemin du fichier csv à importer}
}
\value{
L'object R comportant les données issues de l'export GeoNature
}
\description{
Importer un fichier csv provenant de GeoNature
Une fonction pour importer un fichier csv issu d'un export
GeoNature, sans modification d'encodage ni de format.
}
